requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.0
huggingface-hub==0.20.1
diffusers==0.26.0
torch==2.1.2
transformers==4.36.2
accelerate==0.26.1

